Replace Golden Cookie logic
I went into this project thinking "I'll just replace the function that handles Lucky!, should be easy". I didn't expect every possibly Golden Cookie outcome to be in a single function. Worse, the part to update is right in the middle, so I can't just call the original function with a wrapper.

`newGoldenCookiePop` is copied directly from the original game's code, with one branch of the `else if` changed as noted by the comment of the "novel part".

This means that the mod will break if `popFunc` has any changes across versions, or at least, it will wipe away any version differences. Until the code is more modular, there isn't really a less invasive option (unless there are some extremely clever pro JavaScript things I'm ignorant of).

The whole `Game.shimmerTypes = ` thing from the previous commit technically started from my testing of these changes. I learned that `popFunc` would get executed in the context of the Golden Cookie object, so `this` takes on a different meaning. Initially I was using `Game.mods['persistent lucky bank'].calcLuckyAmount`, but I didn't like the dependency on the string mod ID (which I changed a few times). And it takes up a lot of visual space.

So, I injected the helper functions just like `popFunc` and now `this.helper()` works in both contexts.